♻️ [refactor]: 불필요한 레디스 설정 제거
