+++ src/main/java/com/PickOne/domain/term/dto/TermRequestDto.java
@@ -6,9 +6,6 @@ import jakarta.validation.constraints.NotNull;
 
 import java.time.LocalDateTime;
 
-/**
- * 약관 등록 요청 DTO
- */
 public record TermRequestDto(
         @NotBlank String title,
         @NotBlank String content,
