+++ src/main/java/com/PickOne/global/redis/RedisConfig.java
@@ -10,8 +10,6 @@ import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.listener.ChannelTopic;
-import org.springframework.data.redis.listener.RedisMessageListenerContainer;
 import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
 
@@ -49,22 +47,5 @@ public class RedisConfig {
         redisTemplate.setValueSerializer(serializer);
 
         return redisTemplate;
-    }
-
-    @Bean
-    public RedisMessageListenerContainer redisMessageListener(RedisConnectionFactory connectionFactory) {
-        RedisMessageListenerContainer container = new RedisMessageListenerContainer();
-        container.setConnectionFactory(connectionFactory);
-        return container;
-    }
-
-    @Bean
-    public ChannelTopic messageTopic() {
-        return new ChannelTopic("messaging");
-    }
-
-    @Bean
-    public ChannelTopic notificationTopic() {
-        return new ChannelTopic("notifications");
     }
 }
