--- src/main/java/com/PickOne/global/redis/RedisMessagePublisher.java
@@ -1,25 +1,0 @@
-package com.PickOne.global.redis;
-
-import com.PickOne.domain.messaging.dto.MessageDto;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.redis.core.StringRedisTemplate;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class RedisMessagePublisher {
-
-    private final StringRedisTemplate redisTemplate;
-    private final ObjectMapper objectMapper;
-
-    public void publishMessage(MessageDto messageDto) {
-        try {
-            String json = objectMapper.writeValueAsString(messageDto);
-            redisTemplate.convertAndSend("channel:messaging", json);
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException("메시지 직렬화 실패", e);
-        }
-    }
-}
